Support prefix and exact route at same path
This is sometimes needed when the landing page for a thing is handled by
a separate app from the rest of the thing.

This has involved changing the triemux to store 2 tries, one for the
prefix routes, and one for the exact routes.  This also now uses the
GetLongestPrefix method to do the lookup in the prefixTrie, which means
that each path lookup now involves at most 2 trie lookups.