[#58693908] Bail if unable to bind TCP ports
We should bail early and with an error if we're unable to bind either of the
TCP ports required. This will prevent silent failures if two instances of
the router are started on the same machine, where previously the second
would just sit there never receiving traffic. Should also reduce false
positive errors from CI tests for similar situations.