Add two specs for double slash handling
The first checks that we aren't confusing the behaviour of the Router and
our test backends, both of which are written in Go. The default behaviour of
`ServeMux` is to issue a redirect to the canonicalised path.

The second checks that `/foo//bar` is routed the same as `/foo/bar` and
doesn't fall through to the root prefix.

Both presently fail.