Don't use DefaultServeMux for simple_backend
By passing our own simple handler we cut out `DefaultServeMux`, whose
default behaviour is to issue a redirect if the request path isn't in it's
canonical form:
```
func (mux *ServeMux) Handler(r *Request) (h Handler, pattern string) {
  if r.Method != "CONNECT" {
    if p := cleanPath(r.URL.Path); p != r.URL.Path {
      _, pattern = mux.handler(r.Host, p)
      return RedirectHandler(p, StatusMovedPermanently), pattern
```

We don't want this because it risks confusing the behaviour of the Router
and the test backend. The backend should be dumb and serve up an identifier
regardless.