Vendorise nickstenning/trie
Pull in the source of `trie`, used by `triemux`, to workaround Go's oddities
with dependency versioning. This will ensure that the Router builds against
the correct version and not against Git master/HEAD.

This comes at the small expense of not making it as easy to use as a generic
library from other projects. However if that becomes a big deal in the
future then we can address it then.

Changed import statement in `triemux` accordingly. Cut from:

https://github.com/nickstenning/trie/tree/f14091f4e64174929178166659349a59f7707909