[#59040160] Throughput and concurrency tests
Tests for throughput (requests per second) and concurrency (requests in
flight at a single point in time). These appear to be mostly limited by the
file descriptor ulimit of the process/shell that they are run from.

Because we're launching the following from rspec all in parallel:

- 2x Vegeta attack.
- 2x Vegeta report.
- 2x Simple backend.
- Router.

The combination of the above, not necessarily the router, is also pretty
harsh on system resources such as CPU. So it would be rude to run these from
CI. As such I'm marking them as optional. You can opt into running as
follows:
```
ulimit -n 4096
rspec --tag ulimits
```