Make jenkins.sh work outside of Jenkins
Make the shared Bundler path conditional based on `$JOB_NAME` (which Jenkins
exports) so that the script can be used to run all tests on your own
machine. Strictly the name is wrong, if you can run it outside of Jenkins,
but it follows the convention we've used elsewhere.

NB: The hyphen after the variable name ensures that we don't bomb out on an
unbound variable when not present.