Remove specific error handling for invalid content-length
This handling hasn't worked since the upgrade to Go 1.2 (the tests were
disabled in 14626c9).  This error handling was always brittle as it
relied on regex matching a specific error string from within the
net/http package.

In actualy fact, this error case is not possible to handle easily within
the router, due to the asynchronous nature of proxying HTTP requests.
Specifically, an invalid content-length can't be detected until we get
to the end of the request body.  At this time, the request headers and
most of the body have already been transmitted to the upstream backend
server.  It also could have started transmitting its response.  The only
way to reliably detect this error would be to buffer the entire request
in the router before transmitting to the backend.  This would have
significant implications for both latency and memory usage.