Fix Makefile race condition
Currently, if gom install errors or is aborted, the _vendor directory
will still get its timestamp updated. This means that future make runs
won't attempt to re-run gom install.

This changes the _vendor make target to use a stamp file to avoid this
issue.  It also combines the task for the project importpath into the
main task to avoid needlessly rerunning gom install every time the
to-level directory's timestamp is updated.