Return a 503 when running with no routes.
If the connection to Mongodb fails when the router first starts up, it
will come up with an empty routing table.  Previously this would return
a 404 for all requests.  These would be served to end users and cached
upstream.

Running with an empty routing table is always going to be erroneous, so
this changes it to return a 503's in this situation. This will cause
upstream caches to serve from stale, and additionally won't be cached
anywhere.