Rate limit route reloading.
Although ReloadRoutes attempts to continue serving while the routes are
being reloaded, it is still possible to trigger a DOS by repeatedly
sending a reload - for instance when a large number of routes are being
registered by the content store.

This change implements a rate limit, set by the ROUTER_RELOAD_INTERVAL
env var (default 1s), so that a reload request waits for that interval
before doing the actual reload, and any further requests within that time
do not trigger an extra reload.