Add test for HTTPS backends
Add a test to verify that the router can connect to a HTTPS backend.

For this, add the `ROUTER_TLS_SKIP_VERIFY` environment variable, which
will cause the router to skip verification of backend TLS certificates
when set to a non-empty value.

This option is only intended for use by the tests. We were hesitant to
do this given that disabling TLS verification adjusts the normal
application behaviour, however we decided this is an acceptable
compromise given that we're testing the TLS transport, not the
verification of TLS certificates (though we may later decide to test for
that separately).

Log a warning when `ROUTER_TLS_SKIP_VERIFY` is enabled to make it clear
that it is not intended for production use.