Support special characters in routes.
The database contains these special routes in their %-encoded form, but
the paths we were comparing against were in their decoded form, and
therefore didn't match. This updates our matching code to decode the
paths from the database before comparing against the request URL so that
we're comparing equivelent forms.