Update mgo to release 2015.10.05
Details announced here: http://blog.labix.org/2015/10/05/mgo-r2015-10-05

Of interest to us is the fix for the name resolution. Go 1.5 has changed
the internals of how DNS resolution is done[1], which means that for hosts
with both IPv4 and IPv6 addresses (eg localhost), the IPv6 one is
sometimes returned. Many mongo versions don't listen on IPv6 by default,
so this led to connection errors. The updated driver version resolves
this.

Note: at present, the release hasn't been tagged in git, hence having to
pin to a commit SHA.

[1] https://golang.org/doc/go1.5#net