Add new behaviour for prefix redirects
By default, prefix redirects preserve the path sections after the
prefix and add them to the redirect destination. This isn’t always the
desired behaviour.
This adds a new `prefix_mode` field for redirect routes, set to
`preserve` (the existing behaviour) or `ignore`. If the field isn’t
present in the route document, this will default to `preserve` for
prefix routes.
If the mode is set to `ignore`, the path sections after the prefix will
be discarded and all requests will be redirected to the exact target.