Make router reloading asynchronous
Use a channel based mechanism for actioning router reload requests from
router-api. POST requests to /reload will now result in a message being
sent to a channel which will cause a goroutine to reload the routes.

The channel has a capacity of 1 in order to handle a request arriving
while routes are being reloaded. In this situation, the message will be
“queued” in the channel ready to be read again once the reload for loop
is complete.

The response for /reload is changed to be a 202 to reflect that the
reload is no longer synchronous.