Clarify the CI build process
The CI build process for Go apps is slightly "different," in that the
app needs to be built from within a Go workspace.  This change simply
comments and clarifies what's going on and why.

Functionally, there's nothing new happening in this change, other than
making use of the ability to pass through a binary location to `make`,
rather than copying it out after the build.