optimise docker image size and adds AWS RDS CA
The build docker image was crazy big (nearly 800MB!!) so this switches
to using a multi-stage build and an alpine container to only include the
relevent executable in the image taking it down to a more respectable
12MB.

ENTRYPOINT should be prefered over CMD alone so you don't have a random
`sh` in the process tree and the target executable is exec'd directly.

We also want to be able to use router with a AWS DocumentDB backend
(mongodb compatible) to do this we need to have the AWS provided CA used
to verify tls connections. So this fetches the CA and places it at
/etc/ssl/certs where go expects to find it.