lower test expectations of how fast routes reload
when testing against DocumentDB (which can be a fair bit slower than a
local mongodb instance) we saw race conditions around the expectation
that routes will appear in 50ms.

to avoid this and make the tests less racey we use `Eventually` to wait
for the route to appear with a 1 second cap of time to ensure things are
still practically performant.

The check that the request returns in less than 1ms was also unrealistic
and totally depends on goroutine activity and as such is very flakey
when run with `-race` flag. As a temp fix up this to 5ms, long term I
think it would be ok to drop this timing check entirely and rely on the
1s test timeout to ensure performance is within reason.