fix issue with mgo client when running with race
there is a race between things that the mgo client package does in the
background and way the race detector/ginkgo runs multiepl
integration_tests's in paralel.

move the logic of the init into the BeforeSuite to fail early and up the
SetSocketTimeout to 10mins to avoid it triggering during test runs.