Upgrade to go version 1.12.9
There are no breaking changes. One error output assertion needed to be changed.

However, mgo, the mongodb client library we use, had trouble on the CI
machines connecting to a cluster.

This took a fair bit of time to figure out.
You can run the library in debug mode with the following code:

mylogger := log.New(os.Stderr, , log.LstdFlags)
mgo.SetLogger(mylogger)
mgo.SetDebug(true)
... use the library e.g. mgo.Dial(rt.mongoURL)

This helped raise the following error:
Failed to resolve server address: localhost:27017
Which wasn't happening for the master branch.
For some reason the resolveAddr function in the
library was failing to connect to localhost.

It maybe a problem with a core module e.g. net.DialTimeout,
but that requires some more investigation.

https://trello.com/c/bh2jDTgp/977-bump-versions-of-govuk-golang-apps