Return a 405 if routes do not allow methods
This adds optional methods to routes, allowing us to
specify the routes that should be accepted by a path.

For example, 'GET /' is acceptable but 'DELETE /' is
not. We know this, so we can return a 405 at the router
level, rather than letting our slower applications
handle this request downstream of router.

This requires router-api to set a list of methods
permitted by a path, so the publishing apps will need
to send this information to the publishing-api, which
sends it to content-store, which eventually sends it
to router-api, which then puts the routes into the db
that router reads from.

Router-API should by default set an empty array for all
routes, or nil, which will permit all methods to be
accepted.

Router API should also allow HEAD should a GET method
be allowed.

https://trello.com/c/v9exwjls/1406-spike-implement-request-method-handling-in-router