Add Dependabot config
[RFC 126] suggests configuring Dependabot to only
raise Pull Requests for the following three categories:

1. Security updates
2. Internal libraries
3. Framework libraries

Things have moved on slightly since the RFC; we're now using
[GitHub native] configs rather than the legacy `.dependabot`
approach. The syntax is slightly different to the example config
provided in the RFC but most of it is easily transferable.

There are so few dependencies in Router and Dependabot PRs are
so infrequent that it doesn't seem too important to devise a
specific 'allow' list for Go modules. It's also not trivial to
see what dependencies there are throughout the repository, and we
wouldn't want to miss something crucial.
This commit gives us a blank canvas we can work with - we can
restrict dependencies further in a future commit if needed.

Trello: https://trello.com/c/uPoriyfJ/2049-add-dependabot-configuration-to-each-repo-blitz-pair

[GitHub native]: https://docs.github.com/en/github/administering-a-repository/enabling-and-disabling-version-updates
[RFC 126]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-126-custom-configuration-for-dependabot.md