Report Sentry error when route table is empty
This will report an error to Sentry when a router proc receives
a request when the routing table is empty - currently we don't
have visibility of this.

This shouldn't really happen since we don't want to send user
requests to a router with an empty routing table. We should
use a health check that doesn't raise an error to check if
the routes are loaded.