Use cmd, not entrypoint, in docker image
If an image specifies an ENTRYPOINT, then the CMD or COMMAND
get appended as arguments to the ENTRYPOINT when running
the container.

For example, 'docker run govuk/router /bin/sh -c exit 0'
would start the router app, rather than exiting, as
expected.

Entrypoint is similar to command since it specifies what
executable to run when the container starts. However,
entrypoint is purposefully more difficult to override.

Additionally, since we are bringing up Router in ECS,
it would be good for this image to be similar to the
others we're running. Entrypoint is supported by
ECS though:
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/task_definition_parameters.html#standard_container_definition_params