Increase StartupDelay during deployments
This overrides the default startup delay set in
the Tablecloth library:
https://github.com/alext/tablecloth/blob/9a28c0f651e7faeb01e0381f263d1ad273cfc188/manager.go#L20

Ther Router child process is returning a 503 error during
deploys since it's routing table is still empty when it
receives its first request. We have recently started monitoring
this error case: https://sentry.io/organizations/govuk/issues/2311515746/

My theory is that the routing table is taking longer and
longer to fill as more content is created. 5 seconds may
have been sufficient time to wait for the child process
to get 'hot' in the past, but it now needs a bit longer.

This is a small patch rather than fixing the underlying
issue. The more robust fix will come when we move to
hosting router in ECS, where we'll do blue-green deploys
and remove the Tablecloth dependency.