Standardise the tech docs section
This uses the intro wording we have in other repos, but preserves
the link to Go docs, since we currently have limited expertise on
this within GOV.UK, and the use of Go is surprising / unusual.

As with other repos, GOV.UK codifies the startup instructions [1]
[2], so we don't need to write them here. We can always reinstate
them if there's a need to run the app without GOV.UK Docker.

[1]: https://github.com/alphagov/govuk-docker/blob/master/projects/router/docker-compose.yml#L40
[2]: https://github.com/alphagov/govuk-docker/blob/master/projects/router/Makefile#L2