Minor tweaks to testing documentation
This removes an 8 year old sentence about tests not being run, at
a time when they were written in Ruby [1]. I can't find evidence
that this is still true. We can always add it back and link to the
evidence if we find it in future; for now, it's just confusing.

[1]: https://github.com/alphagov/router/pull/59/files#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R68