WIP: Record triemux lookup duration metrics
WIP because this commit doesn't adequately unit test that the Histogram metric

I think this data would be interesting to look at when talking about
alternative options to the triemux approach.

My hypothesis is that the trie based lookups are extremely fast (low number of
microseconds), since they're operating in memory over a more or less optimal
data structure for the use case.

However, the time taken in the backend handlers is relatively slow (hundreds of
milliseconds), so it might not be wise for us to be chasing after every microsecond.

The current trie based approach does have some significant drawbacks in terms
of complexity of router:
- we have to load the entire database into memory
- we have to poll the database for changes
- we have to hot-reload the in memory copy of the database without dropping requests

Perhaps it would be worth switching to a more conventional database lookup (i.e.
a network call and a B-Tree index scan). This would probably be orders of
magnitude slower than the trie based lookup (largely because of the network
call), but could still be negligible compared to the time taken in the backend handlers.

Having metrics would help us put concrete numbers on some of these things.