Migrate Router functionality to PSQL
Instead of using a polling system, we now use PSQL's Notify channels to inform Router instances that they need to be updated with new routes.

This allows us to move over entirely to Postgres while reducing a lot of the complexity required to ensure we had an up to date status on the router-api database.

Currently we still poll the notify channel every two seconds.