Make the timeout test work without a firewall.
We don't need a firewall, just an address that is always going to be a
black hole. The 240.0.0.0/4 block will do just fine for the foreseeable
future.

We could use v6 (which has a dedicated blackhole address as part of the
standards), but Router doesn't use v6 in production so let's stick with
v4.