Eliminate patch version update toil.
We always want to use the latest patch version of our pinned minor
version of Go by default.

This eliminates the toil of keeping the patch version up-to-date and,
more importantly, eliminates human forgetfulness from the security
update path.

In the extremely unlikely event of a golang patch release causing
a regression, it's still trivial to pin the patch version until it's
fixed.

This also stops us pinning the Go version in 3 different places, which
was getting rather silly. go.mod is now the only place where the Go
version (major.minor) should normally be specified.