Update tsenart/vegeta to current version.
We were stuck on v12.7.0 because we weren't specifying `/v12` in the
import.

Recent-ish versions of Go require explicitly specifying the major
version in the import statement like this when the major version is
greater than 1. This is so that `go get -u` doesn't automatically break
things by updating major versions of libraries.