Avoid automatically registering metrics globally.
Only the main module now touches the default Prometheus registry.

We want to be able to refactor some of our unnecessarily-end-to-end
tests into integration tests (in the true sense) and unit tests. To do
that, we can't have components messing with global state.

This also improves separation of concerns generally. For example, the
triemux and handlers modules can no longer cause panics simply by being
imported if there's a metric name clash.