Migrate tests to reflect new PostgresDB design
Very little of the tests themselves had to be rewritten to reflect
the system being postgres, however a number of tests that relied on
mocking Mongo are no longer valid in the Postgres implementation
and have been removed.

We now spin up a test PSQL database container using an initialiser
sql file to test against.