Add a configurable default backend.
For all requests that don't match any path or path prefix in the
database, forward the request to ROUTER_DEFAULT_BACKEND_URL. This
defaults to `government-frontend`, which is where 93% of routes go.

This should allow shrinking the route database by an order of magnitude
as well as making it a little easier to eliminate Router in the long
run.

We no longer respond 503 to all requests when the route database is
empty. This feature was presumably conceived as a kind of lameduck
mode (albeit somewhat buggy). We're almost certainly better off without
it now that an empty database doesn't mean we're completely broken.