Update min Go version to 1.22.
Use the new version format [introduced] in Go 1.21. This was why CodeQL
wasn't working with 1.22 before: `go 1.22` is not a valid [version]
any more. Versions like `1.20` were always shorthand for the initial
release such as `1.20.0`.

Since the version in go.mod is still a minimum version constraint, we
still don't need to worry about updating the patch version there. It can
just stay as `.0` unless there's some exceptional reason to change it.

[introduced]: https://tip.golang.org/doc/go1.21#introduction
[version]: https://tip.golang.org/doc/toolchain#version