Load backends from env vars
As we only have a small number of backends that change infrequently,
their URL configuration can be handled by environment variables. This
allows us to remove a dependency from using the Mongo Database to fetch
this information and allows us to using the Content Store for route
information.

Router will use any environment variable with the prefix "BACKEND_URL_",
where the key suffix is the backend id and the value is the URL.