Fix periodic reload
Currently, it reloads every minute regardless if a reload has occured
within that minute. This is because we never set the last reload time.
This also improves the responsivness by checking for a periodic reload
every 5 seconds, instead of a minute (which could potentially mean a
reload could take up to 2 minutes).