Allow content store update notification listeners to be disabled
There are scenarios where we would like to be able to load the routes from the
content store database on start, and periodically, but we do not want to listen
for changes in the database, such as during a database migration when the
database briefly becomes read-only.

There are no tests for this behaviour because the pgxmock package that we use
for testing interactions with the database cannot test for listeners being
added. It's mocking features are centerd around the pool type, and the
listeners operate directly on the connection type.