Add an integration test for content store listeners being disabled
New integration test starts the router with the listeners disabled and the
refresh timer set to 1h, adds a new router, and tries repeeatedly to use the
route for 5s. If it succeeds at any point, that is considered a failiure
becuase it should not be being notified of the new router.